revocations byGender, byRace, byRiskLevel utils refactored
