+++ .eslintignore
@@ -73,6 +73,7 @@ public/
 !src/components/charts/new_revocations/RevocationsByRiskLevel/**/*.js
 !src/components/charts/new_revocations/RevocationsByViolation.js
 !src/components/charts/new_revocations/ToggleBar/**/*.js
+!src/components/charts/new_revocations/utils/*.js
 !src/components/charts/programming/FtrReferralCountByMonth.js
 !src/components/charts/programming/FtrReferralsByAge.js
 !src/components/charts/programming/FtrReferralsByGender.js
