+++ .eslintrc.json
@@ -23,7 +23,8 @@
           "some": ["nesting", "id"]
         }
       }
-    ]
+    ],
+    "import/prefer-default-export": "off"
   },
   "env": {
     "browser": true,
