+++ src/components/charts/new_revocations/RevocationsByGender/constants.js
@@ -1,0 +1,6 @@
+import { COLORS } from "../../../../assets/scripts/constants/colors";
+
+export const CHART_COLORS = [
+  COLORS["lantern-light-blue"],
+  COLORS["lantern-orange"],
+];
