+++ src/components/charts/new_revocations/RevocationsByGender/generateRevocationsByGenderChartData.js
@@ -19,8 +19,13 @@ import getOr from "lodash/fp/getOr";
 import pipe from "lodash/fp/pipe";
 import set from "lodash/fp/set";
 import toInteger from "lodash/fp/toInteger";
+import reduce from "lodash/fp/reduce";
+
 import { calculateRate } from "../helpers/rate";
-import { translate } from "../../../../views/tenants/utils/i18nSettings";
+import { getBarBackgroundColor } from "../../../../utils/charts/significantStatistics";
+import { riskLevelLabels } from "../../../../utils/transforms/labels";
+import { CHART_COLORS } from "./constants";
+import getDenominatorKeyByMode from "../utils/getDenominatorKeyByMode";
 
 /**
  * These are the only genders that are apparent in the source data set,
@@ -36,7 +41,7 @@ const RISK_LEVELS = [
   "VERY_HIGH",
 ];
 
-export const getCounts = (transformedData) => {
+const getCounts = (transformedData) => {
   const dataPoints = [];
   const numerators = [];
   const denominators = [];
@@ -60,26 +65,6 @@ export const getCounts = (transformedData) => {
   return { dataPoints, numerators, denominators };
 };
 
-export const getDenominatorKeyByMode = (mode) => {
-  switch (mode) {
-    case "rates":
-    default:
-      return "total_supervision_count";
-    case "exits":
-      return "total_exit_count";
-  }
-};
-
-export const getLabelByMode = (mode) => {
-  switch (mode) {
-    case "rates":
-    default:
-      return translate("percentOfPopulationRevoked");
-    case "exits":
-      return "Percent revoked out of all exits";
-  }
-};
-
 /**
  * Transform to
  * {
@@ -87,7 +72,7 @@ export const getLabelByMode = (mode) => {
  *   MALE: { LOW: [2, 9], HIGH: [2, 8], ... } }
  * }
  */
-export const dataTransformer = (numeratorKey, denominatorKey) => (acc, data) =>
+const dataTransformer = (numeratorKey, denominatorKey) => (acc, data) =>
   pipe(
     set(
       [data.gender, data.risk_level],
@@ -108,3 +93,34 @@ export const dataTransformer = (numeratorKey, denominatorKey) => (acc, data) =>
       ]
     )
   )(acc);
+
+const generateRevocationsByGenderChartData = (
+  apiData,
+  dataFilter,
+  mode,
+  stateCode
+) => {
+  const numeratorKey = "population_count";
+  const denominatorKey = getDenominatorKeyByMode(mode);
+
+  const { dataPoints, numerators, denominators } = pipe(
+    dataFilter,
+    reduce(dataTransformer(numeratorKey, denominatorKey), {}),
+    getCounts
+  )(apiData);
+
+  const generateDataset = (label, index) => ({
+    label,
+    backgroundColor: getBarBackgroundColor(CHART_COLORS[index], denominators),
+    data: dataPoints[index],
+  });
+
+  const data = {
+    labels: riskLevelLabels(stateCode),
+    datasets: ["Women", "Men"].map(generateDataset),
+  };
+
+  return { data, numerators, denominators };
+};
+
+export default generateRevocationsByGenderChartData;
